requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# OpenAI & LLM
openai==1.3.5
tiktoken==0.5.1

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# PDF Processing (with OCR)
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Excel Processing (for chat transcripts)
openpyxl==3.1.2
xlrd==2.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6

# HTTP & Webhooks
httpx==0.25.2
requests==2.31.0

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3
